==> Code/03_baseline_models.Rmd <==
---
title: "03_baseline_models"
output: html_document
date: "`r Sys.Date()`"
editor_options: 
  chunk_output_type: console
---

# Importaciones

```{r}
library(tidyverse)
library(tidymodels)
library(lubridate)
library(here)

theme_set(theme_bw())
theme_update(
  text = element_text(size = 20)
)

data <- readRDS(here("Data", "construccion_clean.rds")) |> 
  mutate(year = year(fecha)) |> 
  filter(year <= 2020)
```

## Crecimiento anual

```{r}
crecimiento <- data |> 
  group_by(accion, year) |> 
  arrange(year) |> 
  summarise(ventas_anuales = sum(ingreso_neto), groups = "drop_last") |> 
  summarise(oportunidades_crec = log(ventas_anuales / dplyr::lag(ventas_anuales)), .groups = "drop") |> 
  mutate(oportunidades_crec = if_else(is.na(oportunidades_crec), 0, oportunidades_crec))

crecimiento <- data |> 
  distinct(accion, year) |> 
  mutate(oportunidades_crec = crecimiento$oportunidades_crec)
```


# Vásquez-Larre 

## Feature engineeirng

```{r}
vas_larre_data <- data |> 
  left_join(crecimiento, by = c("accion", "year")) |> 
  transmute(
    accion,
    fecha,
    endeudamiento_cp = pasivo_corriente / activo_total,
    tamano_empresa = log(activo_total),
    tangibilidad_act = activo_tangible / activo_total,
    oportunidades_crec,
    margen_bruto = resultado_bruto / ingreso_neto,
    p_vl,
    ebitda,
    ciclo_operativo,
    CT_neto = capital_trabajo / activo_total,
    CT_neto_2 = CT_neto ^ 2,
    ROA = utilidad_neta / activo_total
  ) |> 
  na.omit() |> 
  filter(!if_any(where(is.numeric), is.infinite))
```

```{r}
saveRDS(vas_larre_data, here("Data", "vasquez_larre_data.rds"))
```


```{r}
corr_data <- vas_larre_data |> 
  corrr::correlate(diagonal = 1) |> 
  corrr::shave(upper = FALSE) |> 
  pivot_longer(
    cols = -term,
    names_to = "colname",
    values_to = "corr"
  ) |> 
  mutate(
    rowname = fct_inorder(term),
    colname = fct_inorder(colname),
    label = if_else(is.na(corr), "", sprintf("%1.2f", corr))
  )

ggplot(corr_data, aes(rowname, fct_rev(colname), 
                      fill = corr)) +
  geom_tile() +
  geom_text(aes(
    label = label,
    color = abs(corr) < .75
  )) +
  coord_fixed(expand = FALSE) +
  scale_color_manual(
    values = c("white", "black"),
    guide = "none"
  ) +
  scale_fill_distiller(
    palette = "PuOr",
    na.value = "white",
    direction = 1,
    limits = c(-1, 1),
    name = "Correlación\nde Pearson"
  ) +
  labs(x = NULL, y = NULL) +
  theme(text = element_text(size = 16),
        panel.border = element_rect(color = NA, fill = NA),
        legend.position = c(.85, .8),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5),
        axis.ticks = element_blank(),
        panel.grid = element_blank())
```


```{r}
metrica = "CT_neto"
metrica2 = "ROA"

rolling_metrics <- vas_larre_data |> 
  select(fecha, metrica = {{metrica}}, metrica2 = {{metrica2}}, accion) |> 
  mutate(year = year(fecha)) |> 
  group_by(accion) |> 
  mutate(metrica_rolling = (metrica + lag(metrica) + lag(metrica, 2) + lag(metrica, 3)) / 4,
         metrica2_rolling = (metrica2 + lag(metrica2) + lag(metrica2, 2) + lag(metrica2, 3)) / 4) |>
  ungroup() |> 
  pivot_longer(cols = c(metrica, metrica2, metrica_rolling, metrica2_rolling)) |> 
  mutate(name = str_replace(name, "metrica2", metrica2),
         name = str_replace(name, "metrica", metrica)) |> 
  pivot_wider(names_from = name, values_from = value)
  
GGally::ggpairs(rolling_metrics |> select(4:7))
```


## Data partition

```{r}
set.seed(1234)
splits <- initial_split(vas_larre_data)
train_set <- training(splits)
test_set <- testing(splits)

set.seed(2345)
folds <- vfold_cv(train_set, v = 10)
```

## Recipe

```{r}
# Base recipe
base_rec <- recipe(ROA ~ ., data = train_set)
```

## Model specs

```{r}
vas_larre_spec <- linear_reg()
```

## Workflows

```{r}
vas_larre_wf <- workflow() |> 
  add_model(vas_larre_spec) |> 
  add_recipe(base_rec)
```

## Model training

```{r}
vas_larre_res <- fit_resamples(
  vas_larre_wf,
  resamples = folds,
  metrics = metric_set(rmse, rsq),
  control = control_resamples(save_pred = TRUE, verbose = TRUE)
)

vas_larre_res |>
  unnest(.metrics) |> 
  ggplot(aes(.metric, .estimate)) +
  geom_boxplot() +
  annotate(geom = "point", x = "rsq", y = 0.03, size = 5, color = "cyan4") +
  annotate(geom = "text", x = 1.5, y = 0.09, size = 5,
           label = "Modelo de Vásquez y Larre") +
  geom_curve(x = 1.5, xend = 2, y = 0.08, yend = 0.03,
               arrow = arrow(angle = 15, length = unit(0.3, "cm")),
             curvature = 0.2) +
  labs(title = "Resultados de 10-fold cross-validation de regresión lineal",
       x = NULL, y = "valor estimado") 
```
