==> Code/04_random_effects_models.Rmd <==
---
title: "Random effect models"
output: html_document
date: "`r Sys.Date()`"
editor_options: 
  chunk_output_type: inline
---

```{r}
library(tidyverse)
library(lubridate)
library(here)
library(patchwork)

theme_set(theme_bw())
theme_update(
  text = element_text(size = 16),
  plot.title = element_text(hjust = 0.5),
  plot.subtitle = element_text(hjust = 0.5)
)

data <- readRDS(here("Data", "construccion_clean.rds")) |> 
  mutate(year = year(fecha)) |> 
  filter(year <= 2020)

min_max_scaler <- function(x) {
  scaled <- (x - min(x)) / (max(x) - min(x))
  return (scaled)
}

vl_data <- readRDS(here("Data", "vasquez_larre_data.rds")) |> 
  select(-c(ROA, ebitda)) |> 
  mutate(across(c(tamano_empresa, ciclo_operativo), min_max_scaler))

# pib <- readRDS(here("Data", "pib_Chile.rds")) |>
#   mutate(year = as.numeric(year))
```

## Data partitioning

No hay data partitioning, porque el mindset ed este proyecto es prescriptivo, del análisis de los residuales y de la bondad del ajuste.

```{r}
# set.seed(1234)
# train_ids <- sample(nrow(vl_data), size = round(0.7 * nrow(vl_data)), replace = FALSE)
# vl_train <- vl_data[train_ids, ]
# vl_test <- vl_data[-train_ids, ]
```

## Correlation analysis

```{r}
corr_data <- vl_data |>
  select(-c(accion, fecha)) |> 
  relocate(margen_bruto, .after = last_col()) |> 
  corrr::correlate(diagonal = 1) |> 
  corrr::shave(upper = FALSE) |> 
  pivot_longer(
    cols = -term,
    names_to = "colname",
    values_to = "corr"
  ) |> 
  mutate(
    rowname = fct_inorder(term),
    colname = fct_inorder(colname),
    label = if_else(is.na(corr), "", sprintf("%1.2f", corr))
  )

corr_plot <- ggplot(corr_data, aes(rowname, fct_rev(colname), 
                                   fill = corr)) +
  geom_tile() +
  geom_text(aes(
    label = label,
    color = abs(corr) < .75
  )) +
  coord_fixed(expand = FALSE) +
  scale_color_manual(
    values = c("white", "black"),
    guide = "none"
  ) +
  scale_fill_distiller(
    palette = "PuOr",
    na.value = "white",
    direction = 1,
    limits = c(-1, 1),
    name = "Correlación\nde Pearson"
  ) +
  labs(x = NULL, y = NULL,
       title = "Matriz de correlación") +
  theme(text = element_text(size = 16),
        panel.border = element_rect(color = NA, fill = NA),
        legend.position = c(.85, .8),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5),
        axis.ticks = element_blank(),
        panel.grid = element_blank())

ggsave(
  filename = here("Plots", "matriz_predictores.png"),
  plot = corr_plot,
  dpi = 300,
  device = ragg::agg_png
)
```

# Modelo OLS

```{r}
ols <- lm(margen_bruto ~ ., data = vl_data |> select(-c(accion, fecha)))
summary(ols)
```

El modelo **OLS** ajustado a las variables de Vásquez y Larre encuentra efectos significativos para 7 de los 9 predictores, entre ellos $CT$ y $CT^2$. El coeficiente de determinación ajustado $R^2_{adj} = 0.4739$ indica que el ajuste es insuficiente para explicar la mitad de la variable respuesta, margen bruto.

El estadístico $F$ para la prueba de regresión es significativo, para niveles de significancia menores que 0.0000001. 

```{r}
lmtest::bptest(ols)
```


## Análisis de residuales

### Normalidad

```{r}
ols_residuals <- fortify(ols) |> 
  mutate(id = row_number())
n_sturges <- nclass.Sturges(ols_residuals$.resid)

ols_hist <- ols_residuals |> 
  ggplot(aes(x = .resid)) +
  geom_histogram(
    aes(y = ..density.., fill = ..count..), 
    bins = n_sturges
  ) +
  stat_function(
    fun = dnorm, 
    color = "firebrick",
    size = 1,
    args = list(mean = mean(ols_residuals$.resid),
                sd = sd(ols_residuals$.resid))
  ) +
  scale_fill_gradient(low = "#dcdcdc", high = "#7c7c7c") +
  labs(
    title = "Histograma de residuales\ny curva normal teórica",
    x = "residuales", 
    y = "densidad",
    fill = "frecuencia"
  )

ols_qq <- ols_residuals |> 
  ggplot(aes(sample = .stdresid)) +
  geom_qq_line(
    color = "firebrick", 
    size = 1,
  ) +
  geom_qq(
    alpha = 0.4,
    shape = 16,
    size = 2,
  ) +
  labs(
    title = "Gráfico Normal Cuantil-Cuantil",
    x = "cuantiles teóricos",
    y = "cuantiles muestrales"
  )

ols_plot_1 <- ols_hist + ols_qq + plot_annotation(tag_levels = "a", tag_suffix = ")")
ggsave(
  filename = here("Plots", "OLS_normalidad.png"),
  plot = ols_plot_1,
  device = ragg::agg_png,
  dpi = 300,
  height = 7,
  width = 11
)
```

De acuerdo con el histograma, los residuales tienen una distribución aproximadamente normal, con sesgo positivo, lo que se evidencia con el gráfico cuantil-cuantil. La cola derecha de la distribución tiene un outlier, el que podría ser influyente en el modelo OLS. Para complementar el análisis de normalidad, se ejecuta la prueba no paramétrica de bondad de ajuste de Kolmogorov-Smirnov, con un nivel de significancia del 5%.

```{r}
ks.test(
  x = ols_residuals$.resid, 
  y = pnorm,
  mean = mean(ols_residuals$.resid),
  sd = sd(ols_residuals$.resid)
)
```

Se observa un resultado significativo para el estadístico $D$ del test de Kolmogorov-Smirnov. Por tanto, se concluye que la distribución de los residuales del modelo OLS no siguen una distribución Normal. 

```{r}
ggplot(
  data = ols_residuals,
  aes(x = .fitted, y = .resid)
) +
  geom_point(
    alpha = 0.4,
    shape = 16,
    size = 2
  ) +
  geom_smooth(
    method = "loess",
    formula = y ~ x,
    se = FALSE,
    color = "firebrick",
    span = 0.75
  ) +
  labs(
    title = "Residuales vs valores ajustados",
    x = "valores ajustados",
    y = "residuales"
  )
```

```{r}
ols_scale_plot <- ggplot(
  data = ols_residuals,
  aes(x = .fitted, y = sqrt(abs(.stdresid)))
) +
  geom_point(
    alpha = 0.4,
    shape = 16,
    size = 2
  ) +
  geom_smooth(
    method = "loess",
    formula = y ~ x,
    se = FALSE,
    color = "firebrick",
    span = 0.75
  ) +
  labs(
    title = "Escala-ubicación",
    x = "valores ajustados",
    y = expression(sqrt(abs(residuales~estandarizados)))
  )

cd_cont_pos <- function(leverage, level, model) {
  sqrt(level * length(coef(model)) * (1 - leverage) / leverage)
}

cd_cont_neg <- function(leverage, level, model) {
  -cd_cont_pos(leverage, level, model)
}

cd_labels <- tibble(
  x = rep(c(0.35, 0.4), 2),
  y = c(3.6, 4.4, -3.6, -4.4),
  label = rep(c(0.5, 1), 2)
)

ols_cook_plot <- ggplot(
  data = ols_residuals,
  aes(x = .hat, y = .stdresid)
) +
  geom_hline(yintercept = 0, linetype = 2) +
  geom_vline(xintercept = 0, linetype = 2) +
  geom_point(
    alpha = 0.4,
    shape = 16,
    size = 2
  ) +
  stat_function(
    fun = cd_cont_pos,
    args = list(level = 0.5, model = ols),
    linetype = 2,
    color = "forestgreen"
  ) +
  stat_function(
    fun = cd_cont_pos,
    args = list(level = 1, model = ols),
    linetype = 2,
    color = "forestgreen"
  ) +
  stat_function(
    fun = cd_cont_neg,
    args = list(level = 0.5, model = ols),
    linetype = 2,
    color = "forestgreen"
  ) +
  stat_function(
    fun = cd_cont_neg,
    args = list(level = 1, model = ols),
    color = "forestgreen",
    aes(linetype = "distancia de Cook")
  ) +
  scale_y_continuous(limits = c(-5, 5)) +
  geom_smooth(
    method = "loess",
    formula = y ~ x,
    se = FALSE,
    color = "firebrick",
    span = 0.75
  ) +
  geom_text(
    data = cd_labels,
    aes(x = x, y = y, label = label),
    color = "forestgreen"
  ) +
  labs(
    title = "Residuales vs levarage",
    x = "leverage",
    y = "residuales estandarizados"
  ) +
  scale_linetype_manual(
    name = NULL, 
    values = c(2)
  ) +
  theme(legend.position = c(0.3, 0.15),
        legend.background = element_rect(fill = "transparent"),
        legend.key = element_rect(fill = "transparent"))

ols_plot_2 <- ols_scale_plot + ols_cook_plot + plot_annotation(tag_levels = "a", tag_suffix = ")")

ggsave(
  filename = here("Plots", "Scale_Cook_plot.png"),
  plot = ols_plot_2,
  device = ragg::agg_png,
  dpi = 300,
  height = 7,
  width = 11
)
```

```{r}
ggplot(
  data = ols_residuals,
  aes(x = id, y = .cooksd)
) +
  geom_col() +
  labs(
    x = "número de observación",
    y = "distancia de Cook",
    title = "Distancia de Cook"
  )
```

Claramente, hay una observación influyente en el modelo OLS. La observación mencionada corresponde al cuarto trimestre de 2019 de la empresa Inmobiliaria Apoquindo S.A., período que coincide con su más alto endeudamiento de corto plazo. Además, la empresa no declara sus activos tangibles para el conjunto de entrenamiento, lo que implica que el capital de trabajo para todos sus períodos es negativo. Adicionalmente, los costos de ventas son 0, evidencia de esto es el margen bruto igual a la unidad. En síntesis, la empresa no declara costos de ventas ni activos tangibles para el período de estudio.

# ME model

```{r}
library(lme4)

vl_re_data <- vl_data |>
  group_by(accion) |> 
  mutate(time = 1:n() - 1) |> 
  ungroup() 

me <- lmerTest::lmer(
  margen_bruto ~ endeudamiento_cp + tamano_empresa + 
    tangibilidad_act +  oportunidades_crec +
    p_vl + ciclo_operativo + CT_neto +
    CT_neto_2 + (1 | accion),
  data = vl_re_data
)

broom.mixed::tidy(me, conf.int = TRUE, conf.level = 0.95) |> select(effect, term, estimate, p.value)
summary(me)
0.288856 / 0.593386
```

```{r}
me_slope <- lmerTest::lmer(
  margen_bruto ~ endeudamiento_cp + tamano_empresa + 
    tangibilidad_act +  oportunidades_crec +
    p_vl + ciclo_operativo + CT_neto +
    CT_neto_2 + (time | accion),
  data = vl_re_data
)

me_residuals <- me |> broom::augment()
me_slope_residuals <- me_slope |> broom::augment()
```


```{r}
map(
  list(me_residuals, me_slope_residuals),
  \(x) (ks.test(
    x = x$.resid, 
    y = pnorm,
    mean = mean(x$.resid),
    sd = sd(x$.resid)
  ))
)
```


```{r}
anova(me, me_slope)
```

```{r}
ranef(me)
ranef(me_slope)
```

```{r}
MuMIn::r.squaredGLMM(me)
MuMIn::r.squaredGLMM(me_slope)
```

```{r}
summary(me_slope)
0.440196 / 0.430534
```

```{r}
me_resid_plot <- ggplot(
  data = me_residuals,
  aes(x = .fitted, y = .resid)
) +
  geom_point(
    aes(color = .cooksd >= 0.4),
    show.legend = FALSE,
    alpha = 0.4,
    shape = 16,
    size = 2
  ) +
  geom_smooth(
    method = "loess",
    formula = y ~ x,
    color = "firebrick",
    se = FALSE,
    size = 1
  ) +
  scale_color_manual(values = c("black", "firebrick")) +
  labs(
    x = expression(hat(y)),
    y = "residuales",
    title = "Valores ajustados vs residuales"
  )


me_cook_plot <- ggplot(
  data = me_residuals,
  aes(x = seq_along(.cooksd), y = .cooksd)
) +
  geom_col(aes(color = .cooksd >= 0.4), show.legend = FALSE) +
  scale_color_manual(values = c("black", "firebrick")) +
  labs(
    x = "número de observación",
    y = "distancia de Cook",
    title = "Distancia de Cook"
  )

me_plot_1 <- me_resid_plot + me_cook_plot + plot_annotation(tag_levels = "a", tag_suffix = ")")


ggsave(
  filename = here("Plots", "LMER_residuales_Cook.png"),
  plot = me_plot_1,
  device = ragg::agg_png,
  dpi = 300,
  height = 7,
  width = 11
)
```


```{r}
ids_to_remove <- me_residuals |> 
  mutate(id = row_number()) |> 
  slice_max(abs(.resid), n = 7) |> 
  filter(margen_bruto == 1) |> 
  pull(id)

# vl_data_pib <- vl_data |> 
#   mutate(year = year(fecha)) |> 
#   left_join(pib |> mutate(pib_lag = lag(pib)), by = c("year")) |> 
#   mutate(pib = min_max_scaler(pib),
#          pib_lag = min_max_scaler(pib_lag))

me_2 <- lmerTest::lmer(
  margen_bruto ~ endeudamiento_cp + tamano_empresa + 
    tangibilidad_act +  oportunidades_crec +
    p_vl + ciclo_operativo + CT_neto +
    CT_neto_2 + (1 | accion),
  data = vl_data[-ids_to_remove, ]
)

me_2_residuals <- broom::augment(me_2)

MuMIn::r.squaredGLMM(me_2)
summary(me_2)
0.259268 / 0.62079
```

```{r}
me_2_resid_plot <- ggplot(
  data = me_2_residuals,
  aes(x = .fitted, y = .resid)
) +
  geom_point(
    aes(color = .cooksd >= 0.4),
    show.legend = FALSE,
    alpha = 0.4,
    shape = 16,
    size = 2
  ) +
  geom_smooth(
    method = "loess",
    formula = y ~ x,
    color = "firebrick",
    se = FALSE,
    size = 1
  ) +
  scale_color_manual(values = c("black", "firebrick")) +
  labs(
    x = expression(hat(y)),
    y = "residuales",
    title = "Valores ajustados vs residuales"
  )


me_2_cook_plot <- ggplot(
  data = me_2_residuals,
  aes(x = seq_along(.cooksd), y = .cooksd)
) +
  geom_col(aes(color = .cooksd >= 0.4), show.legend = FALSE) +
  scale_color_manual(values = c("black", "firebrick")) +
  labs(
    x = "número de observación",
    y = "distancia de Cook",
    title = "Distancia de Cook"
  )

me_plot_2 <- me_resid_plot + me_cook_plot + plot_annotation(tag_levels = "a", tag_suffix = ")")
```


```{r}
me_2 <- lmerTest::lmer(
  margen_bruto ~ endeudamiento_cp + tamano_empresa + 
    tangibilidad_act +  oportunidades_crec +
    p_vl + ebitda + ciclo_operativo + CT_neto +
    CT_neto_2 + pib + (1 | accion),
  data = vl_data_pib
)

summary(me_2)
```

```{r}
vl_coeffs <- tibble(
  term = factor(c("endeudamiento_cp", "tamano_empresa", "tangibilidad_act",
                  "oportunidades_crec", "p_vl", "CT_neto", "CT_neto_2")),
  estimate = c(-0.100, 0.004, 0.002, 0.0213, 0.005, 0.048, NA)
)

me_2_coefs_plot <- broom.mixed::tidy(me_2, conf.int = TRUE, conf.level = 0.95) |> 
  filter(effect == "fixed", term != "(Intercept)") |> 
  mutate(term = fct_rev(fct_inorder(term))) |> 
  ggplot(aes(
    x = term, 
    y = estimate, 
    ymin = conf.low, 
    ymax = conf.high,
    color = 0 >= conf.low & 0 <= conf.high
  )) +
  geom_hline(
    yintercept = 0,
    linetype = 2,
    color = "firebrick"
  ) +
  geom_point(
    shape = 16,
    size = 3
  ) +
  geom_errorbar(
    width = 0.4,
    size = 0.8
  ) +
  geom_point(
    data = vl_coeffs,
    aes(x = term, y = estimate),
    inherit.aes = FALSE,
    shape = 16,
    size = 3,
    color = "dodgerblue4"
  ) +
  annotate(
    geom = "text",
    label = "Coeficientes de Vásquez\ny Larre (2020)",
    x = 5,
    y = 0.15,
    hjust = 0,
    lineheight = 0.8
  ) +
  annotate(
    geom = "text",
    label = "Coeficientes del Modelo 4\nde efectos mixtos",
    x = 2.1,
    y = 0.16,
    hjust = 0, 
    lineheight = 0.8
  ) +
  geom_segment(
    aes(x = 5, xend = 5, y = 0.0213, yend = 0.13),
    arrow = arrow(angle = 20, length = unit(3, "mm")),
    color = "dodgerblue4",
    size = 0.5
  ) +
  geom_segment(
    aes(x = 1, xend = 1.8, y = 0.31, yend = 0.31),
    arrow = arrow(angle = 20, length = unit(3, "mm")),
    color = "firebrick",
    size = 0.5
  ) +
  scale_color_manual(values = c("firebrick", "gray")) +
  coord_flip() +
  labs(
    title = "LMER: efectos sobre el margen bruto",
    subtitle = "Intervalo de confianza del 95%",
    y = expression(hat(beta)~"± Margen de error"),
    x = NULL
  ) +
  theme(legend.position = "none")

ggsave(
  filename = here("Plots", "LMER_coefficients.png"),
  plot = me_2_coefs_plot,
  device = ragg::agg_png,
  dpi = 300
)
```
